pyproject.toml
[project]
name = "caad-sendai-engineer-internship-202409"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "caad-suzuki-yutaka", email = "suzuki_yutaka@ca-adv.co.jp" }
]
dependencies = [
    "aiosqlite==0.20.0",
    "chardet==5.2.0",
    "charset-normalizer==3.3.2",
    "duckduckgo-search==6.2.6",
    "japanize-matplotlib==1.1.3",
    "langchain-chroma==0.1.2",
    "langchain-community==0.2.11",
    "langchain-core==0.2.28",
    "langchain-experimental==0.0.64",
    "langchain-openai==0.1.20",
    "langchain==0.2.12",
    "langgraph-checkpoint==1.0.2",
    "langgraph==0.2.3",
    "markdown==3.6",
    "matplotlib==3.9.1",
    "pandas==2.2.2",
    "psutil==6.0.0",
    "python-dotenv==1.0.1",
    "python-magic==0.4.27",
    "python-multipart==0.0.9",
    "python-on-whales==0.72.0",
    "unstructured==0.14.9",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
